Split GradleBuildConfigurationCacheSmokeTest and increase maxParallelForks
So they can be executed in parallel
