+++ subprojects/smoke-test/build.gradle.kts
@@ -84,7 +84,7 @@ tasks {
         group = "Verification"
         testClassesDirs = smokeTestSourceSet.output.classesDirs
         classpath = smokeTestSourceSet.runtimeClasspath
-        maxParallelForks = 1 // those tests are pretty expensive, we shouldn't execute them concurrently
+        maxParallelForks = 2 // those tests are pretty expensive, we should limit the parallelism
         inputs.property("androidHomeIsSet", System.getenv("ANDROID_HOME") != null)
         inputs.property("androidSdkRootIsSet", System.getenv("ANDROID_SDK_ROOT") != null)
         inputs.files(remoteProjects.map { it.map { it.outputDirectory } })
